Moved gather_files and fixed
