+++ conans/test/unittests/util/files/tar_extract_test.py
@@ -8,9 +8,8 @@ import unittest
 import pytest
 
 from conans.client.tools.files import chdir
-from conans.model.manifest import gather_files
 from conans.test.utils.test_files import temp_folder
-from conans.util.files import tar_extract, gzopen_without_timestamps, save
+from conans.util.files import tar_extract, gzopen_without_timestamps, save, gather_files
 
 
 class TarExtractTest(unittest.TestCase):
